Posts card share UI and Functionality added

+++ src/Components/PostCard/PostCard.jsx
@@ -34,7 +34,6 @@ function PostCard({title,image,imagePath,id,author,time,video,activity}) {
     const handleSetProfilePic=()=>{
         const payload ={
             image,
-            imagePath,
             activity: `changed his profile picture.` ,
             author : uid,
             time : new Date()
@@ -45,12 +44,39 @@ function PostCard({title,image,imagePath,id,author,time,video,activity}) {
         
 
     }
+    const handleFav=()=>{
+        database.collection("users").doc(uid).collection("favorites").doc(id).set({postId:id,time:new Date()});
+    }
    
 
     const handleDeletePost=()=>{
-        database.collection("posts").doc(id).delete();
-        storage.ref("postImages").child(imagePath).delete();
         
+        database.collection("posts").doc(id).delete();
+        if(imagePath){
+            storage.ref("postImages").child(imagePath).delete();
+        }
+        
+        
+    }
+
+    const handleShare=(title)=>{
+        let payload = {
+            originalAuthor:author,
+            author:uid,
+            activity:`shared ${userData.first_name} ${userData.last_name}'s post.`,
+            title:title||"",
+            time:new Date()
+        }
+        if(image){
+            payload.image=image;
+        }else if(video){
+            payload.video=video;
+        }
+        if(author===uid){
+            payload.activity=`shared his post.`
+        }
+       database.collection("posts").add(payload);
+
     }
 
     const handleClosePostModal=()=>{
@@ -85,51 +111,61 @@ function PostCard({title,image,imagePath,id,author,time,video,activity}) {
     }
 
     React.useEffect(()=>{
-        database.collection("posts").doc(id).collection("likes").onSnapshot(response=>{
+        const unsubscribe1 = database.collection("posts").doc(id).collection("likes").onSnapshot(response=>{
             setLikes(response.docs.map(doc=>({likeId:doc.id,...doc.data()})))
         })
-    },[])
-
-    React.useEffect(()=>{
-        database.collection("posts").doc(id).collection("comments").orderBy("time","asc").onSnapshot(response=>{
+   
+        const unsubscribe2 = database.collection("posts").doc(id).collection("comments").orderBy("time","asc").onSnapshot(response=>{
             setComments(response.docs.map(doc=>({commentId:doc.id,...doc.data()})))
         })
-    },[])
-
    
-    React.useEffect(() => {
-        const unsubscribe = database.collection("users").doc(author)
+        const unsubscribe3 = database.collection("users").doc(author)
         .onSnapshot((doc) => {
             setUserData(doc.data());
             
         });
         return () => {
-            unsubscribe();
+            unsubscribe1();
+            unsubscribe2();
+            unsubscribe3();
         }
     }, [])
+
+    React.useEffect(() => {
+        if(!image && userData){
+            setLoading(false);
+        }
+    },[userData]);
 
 
 
     return (
         <>
         <div className="postCardContainer" style={{display:loading||!userData?"none":"block"}}>
-            <PostCardHead {...userData} postEditFunction={{handleEditPost,handleDeletePost,handleSetProfilePic}} time={time} author={author} image={image}title={title} activity={activity}/>
+            <PostCardHead handleFav={handleFav} {...userData} postEditFunction={{handleEditPost,handleDeletePost,handleSetProfilePic}} time={time} author={author} image={image}title={title} activity={activity}/>
             {title && <div className="postCardTags">{title}</div>}
-            {image&&<div onClick={()=>setPostModalVisibility(true)} className="postCardImage"><img onLoad={()=>setLoading(false)} src={image|| process.env.PUBLIC_URL + '/Images/facebook_login_logo.png'} alt="img" /></div>}
+            {image&&<div onClick={()=>setPostModalVisibility(true)} className="postCardImage">
+                <img onLoad={()=>setLoading(false)} src={image|| process.env.PUBLIC_URL + '/Images/facebook_login_logo.png'} alt="img" />
+            </div>}
             {video&&<div className="postCardImage">
                 <video width="100%" height="500" controls >
                     <source src={video} type="video/mp4"/>
                 </video>
             </div>}
             <div className="postCardLike flexBox">
-                <div className="flexBox"><LikeEmoji/> <p>{likes.length}</p></div>
-                <div className="flexBox"><p onClick={showComment}>{comments.length} Comments</p> </div>
+                <div className="flexBox">
+                    <LikeEmoji/> 
+                    <p>{likes.length}</p>
+                </div>
+                <div className="flexBox">
+                    <p onClick={showComment}>{comments.length} Comments</p> 
+                </div>
             </div>
-            <PostCardFooter handleDeleteLike={handleDeleteLike} handleLike={handleLike} like={JSON.stringify(likes).includes(uid)} showComment={showComment}/>
+            <PostCardFooter handleDeleteLike={handleDeleteLike} handleLike={handleLike} handleShare={handleShare} like={JSON.stringify(likes).includes(uid)} showComment={showComment} author={author} {...userData} title={title} handleShare={handleShare} image={image}/>
            {commentSection && <PostCardComment postId={id} comments={comments} userData={userData}/>}
            
         </div>
-        {postModalVisibility &&<PostModal handleClosePostModal={handleClosePostModal} uid={uid} image={image} video={video} time={time} userData={userData}  activity={activity} id={id} likes={likes} comments={comments} title={title} handleLike={handleLike} handleDeleteLike={handleDeleteLike} />}
+        {postModalVisibility &&<PostModal handleClosePostModal={handleClosePostModal} uid={uid} image={image} video={video} author={author} time={time} userData={userData}  activity={activity} id={id} likes={likes} comments={comments} title={title} handleLike={handleLike} handleDeleteLike={handleDeleteLike} showComment={showComment} postEditFunction={{handleEditPost,handleDeletePost,handleSetProfilePic}}/>}
        
     
         <div className="postCardContainer" style={{display:loading||!userData?"block":"none"}}>
